equals-space-missing-validator All functionality has been completed and the tests have been carried out. There is now a test folder containing files to use for tests.
