+++ Testing/Test Files/ESMFail1.js
@@ -1,0 +1,5 @@
+class EqualsSpaceMissingFail {
+    pass() {
+        let x=1;
+    }
+}
