+++ Testing/Test Files/ESMFail2.js
@@ -1,0 +1,5 @@
+class EqualsSpaceMissingFail {
+    pass() {
+        let x=1, y=2;
+    }
+}
