+++ Testing/Test Files/ESMFail3.js
@@ -1,0 +1,6 @@
+class EqualsSpaceMissingFail {
+    pass() {
+        let x=1;
+        let y=2;
+    }
+}
