+++ Testing/Test Files/ESMFail5.js
@@ -1,0 +1,6 @@
+class EqualsSpaceMissingFail {
+    pass() {
+        let x =1;
+        let y = 2;
+    }
+}
