+++ Testing/Test Files/ESMPass.js
@@ -1,0 +1,5 @@
+class EqualsSpaceMissingPass {
+    pass() {
+        let x = 1;
+    }
+}
